requirements.txt
biopython
numpy
matplotlib